Added jQuery library
The jQuery library gives us a great range of utilities for easier
handling of documents.
Some of the functionalities are duplicated in D3, but some are not. In
a real world project we would reconsider if jQuery is strictly
necessary, but for making this class as easy as possible we will just
include it.